Fixed many many bugs - Connection is now able to be established.
