+++ app/src/main/AndroidManifest.xml
@@ -52,21 +52,23 @@
             android:label="@string/title_activity_build_info"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".ShowCodeActivity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
 
         <service android:name=".summoner.ADBService" />
 
         <activity android:name=".ConnectionSetupActivity" />
         <activity android:name=".NearbyConnectionActivity" />
-        <activity android:name=".ShowCodeConnAuth" />
+        <activity android:name=".ShowCodeConnAuth">
+        <intent-filter>
+            <action android:name="android.intent.action.MAIN" />
+
+            <category android:name="android.intent.category.LAUNCHER" />
+        </intent-filter>
+        </activity>
         <activity
             android:name=".ClientConnAuthActivity"
-            android:label="@string/title_activity_client_conn_auth"></activity>
+            android:label="@string/title_activity_client_conn_auth">
+        </activity>
     </application>
 
 </manifest>
