+++ app/src/main/java/com/amos/flyinn/ShowCodeConnAuth.java
@@ -3,6 +3,7 @@ package com.amos.flyinn;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.util.Log;
 import android.widget.TextView;
 
 import java.security.SecureRandom;
@@ -17,13 +18,10 @@ import java.security.SecureRandom;
 public class ShowCodeConnAuth extends ClientConnAuthActivity {
 
     /** */
-    private final int CODE_LENGTH = 4;
 
-    /** */
-    private final String appCode = generateNumber(CODE_LENGTH);
 
     private TextView display;
-
+    private static final String CLIENT_NAME_TAG = "Client name is:";
     /**
      *
      * @param savedInstanceState
@@ -32,30 +30,9 @@ public class ShowCodeConnAuth extends ClientConnAuthActivity {
     protected void onCreate(Bundle savedInstanceState) {
         setContentView(R.layout.activity_show_code);
         display = findViewById(R.id.textView2);
-        display.setText(appCode);
+        display.setText(super.appCode);
+        Log.i(CLIENT_NAME_TAG, "Started advertising " + super.clientName);
         super.onCreate(savedInstanceState);
     }
 
-    /**
-     *
-     * @param length
-     * @return
-     */
-    private String generateNumber(int length) {
-        SecureRandom rnd = new SecureRandom();
-        String number = "";
-        for (int i = 0; i < length; i++) {
-            number += rnd.nextInt(10);
-        }
-        return number;
-    }
-
-    /**
-     *
-     * @return
-     */
-    @Override
-    protected String generateName() {
-        return R.string.flyinn_name + appCode;
-    }
 }
