+++ app/src/main/res/values/strings.xml
@@ -12,8 +12,7 @@
     <string name="nearby_connection_rejected">Connection was rejected.</string>
     <string name="nearby_connection_error">Failed to establish connection.</string>
     <string name="nearby_disconnected">Connection closed.</string>
-
-    <string name="flyinn_name">FlyInn-</string>
+    
 
     <string name="buildinfo_git_tag_name">Git Tag: </string>
     <string name="buildinfo_git_commit_date">Git Commit: </string>
