+++ server/src/main/AndroidManifest.xml
@@ -44,11 +44,6 @@
             android:label="@string/title_activity_build_info"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity android:name=".ConnectToClientActivity">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <activity
             android:name=".NearbyServerActivity"
@@ -56,7 +51,13 @@
         <activity
             android:name=".ConnectionSetupServerActivity"
             android:theme="@android:style/Theme.NoTitleBar.Fullscreen" />
-        <activity android:name=".ServerConnectionConnAuth" />
+        <activity android:name=".ServerConnectionConnAuth">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
 
         <activity android:name=".ServerConnAuthActivity"></activity>
     </application>
